API通信部分完成
